feat: Nuxi Devtools command
